+++ src-test/Main.hs
@@ -3,7 +3,7 @@
 
 -- |
 -- Copyright: © Herbert Valerio Riedel 2018
--- SPDX-License-Identifier: GPL-3.0
+-- SPDX-License-Identifier: GPL-2.0-or-later
 --
 module Main where