+++ src/Data/YAML.hs
@@ -4,7 +4,7 @@
 
 -- |
 -- Copyright: © Herbert Valerio Riedel 2015-2018
--- SPDX-License-Identifier: GPL-3.0
+-- SPDX-License-Identifier: GPL-2.0-or-later
 --
 -- Document oriented [YAML](http://yaml.org/spec/1.2/spec.html) parsing API inspired by [aeson](http://hackage.haskell.org/package/aeson).
 --