+++ src/Data/YAML/Loader.hs
@@ -7,7 +7,7 @@
 
 -- |
 -- Copyright: © Herbert Valerio Riedel 2015-2018
--- SPDX-License-Identifier: GPL-3.0
+-- SPDX-License-Identifier: GPL-2.0-or-later
 --
 module Data.YAML.Loader
     ( decodeLoader