+++ src/Data/YAML/Token/Encoding.hs
@@ -4,7 +4,7 @@
 -- |
 -- Copyright: © Oren Ben-Kiki 2007,
 --            © Herbert Valerio Riedel 2015-2018
--- SPDX-License-Identifier: GPL-3.0
+-- SPDX-License-Identifier: GPL-2.0-or-later
 --
 -- UTF decoding
 --